Tests passed

+++ src/Server/ServerStrategySolveSearchProblem.java
@@ -1,25 +1,20 @@
 package Server;
 
-import algorithms.mazeGenerators.IMazeGenerator;
 import algorithms.mazeGenerators.Maze;
-import algorithms.mazeGenerators.Position;
 import algorithms.search.*;
-//import sun.awt.Mutex;
-
 import java.io.*;
 import java.util.concurrent.locks.ReentrantLock;
 
 public class ServerStrategySolveSearchProblem implements IServerStrategy {
 
-    //private Mutex mutex;
-    //private ReentrantLock rLock;
+    private ReentrantLock mutex;
 
     public ServerStrategySolveSearchProblem(){
-        //mutex = new Mutex();
-        //rLock = new ReentrantLock();
+        mutex = new ReentrantLock();
     }
 
     private static String tempDirectoryPath = System.getProperty("java.io.tmpdir");
+
     @Override
     public void serverStrategy(InputStream inFromClient, OutputStream outToClient) {
 
@@ -31,9 +26,8 @@ public class ServerStrategySolveSearchProblem implements IServerStrategy {
             try {
                 Maze maze = (Maze) fromClient.readObject();
                 String tempPath = tempDirectoryPath + "maze" + maze.toString().hashCode();
-                System.out.println(tempPath);
-                //rLock.lock();
-                //mutex.lock();
+
+                mutex.lock();
                 File f = new File(tempPath);
                 if(f.exists()){ // the file exists, we don't need to solve again. only take what exists.
                     FileInputStream fin = new FileInputStream(tempPath);
@@ -41,22 +35,26 @@ public class ServerStrategySolveSearchProblem implements IServerStrategy {
                     solution = (Solution) oin.readObject();
                     fin.close();
                     oin.close();
+                    mutex.unlock();
                 }
                 else{
+                    mutex.unlock();
                     SearchableMaze searchableMaze = new SearchableMaze(maze);
                     solution = Configurations.getAlgorithms_solveAlgorithm().solve(searchableMaze);
-                    f.createNewFile();
-                    FileOutputStream fout = new FileOutputStream(tempPath);
-                    fout.flush();
-                    ObjectOutputStream oout = new ObjectOutputStream(fout);
-                    oout.flush();
-                    oout.writeObject(solution);
-                    oout.flush();
-                    fout.close();
-                    oout.close();
+                    mutex.lock();
+                    if(!f.exists()) {
+                        f.createNewFile();
+                        FileOutputStream fout = new FileOutputStream(tempPath);
+                        fout.flush();
+                        ObjectOutputStream oout = new ObjectOutputStream(fout);
+                        oout.flush();
+                        oout.writeObject(solution);
+                        oout.flush();
+                        fout.close();
+                        oout.close();
+                    }
+                    mutex.unlock();
                 }
-                //mutex.unlock();
-                //rLock.unlock();
                 toClient.writeObject(solution);
                 toClient.flush();
             }
